[BUGFIX] Fix perses dependency by separating ui module
