+++ ui/go.mod
@@ -1,0 +1,3 @@
+module github.com/perses/perses/ui
+
+go 1.18
